GH-22: Fix zwave_command_class_node_info_resolver unit test
GH-22 add a new callback on the give_up_listener on the ATTRIBUTE_ZWAVE_SECURE_NIF attribute causing the test case to fail.

The test only expect this listener to be called on the non-secure NIF ATTRIBUTE_ZWAVE_NIF. This change make sure that the code is only executed in the context of ATTRIBUTE_ZWAVE_NIF and ignored  for ATTRIBUTE_ZWAVE_SECURE_NIF as this test don't care about it.

Bug-SiliconLabs: UIC-3082
Relate-to: https://github.com/SiliconLabs/UnifySDK/pull/22
Forwarded-SiliconLabs: task/UIC-3082/phcoval/GH-22/develop